requirements.txt
face-recognition-models==0.3.0
numpy==2.0.2
opencv-python==4.11.0.86
pillow==11.1.0
